Tests/LibWeb/Text/expected/WebAnimations/misc/get-animations-should-make-sure-style-is-updated.txt

This is a animated div  animation count: 1
